Add OPTION BASE 0.5 for half-based array indexing
Support 0.5 as a valid base for OPTION BASE and OPTION GPU BASE,
enabling tongue-in-cheek half-indexed arrays where A(0.5) is the first
element. Uses runtime arithmetic (int)(index - 0.5) which is exact in
IEEE 754. GPU intrinsics offset by +0.5 when gpu_base is 0.5.

+++ src/ast.h
@@ -204,7 +204,7 @@ struct GpuKernelStmt {
 
 struct OptionBaseStmt {
     bool is_gpu;   // true for OPTION GPU BASE, false for OPTION BASE
-    int base;      // 0 or 1
+    double base;   // 0, 0.5, or 1
 };
 
 struct GpuGosubStmt {
